pyproject.toml
[project]
name = "yaci-s3"
version = "0.1.0"
description = "S3 upload tool for yaci-store Cardano parquet exports"
requires-python = ">=3.10"
dependencies = [
    "click>=8.1",
    "python-dotenv>=1.0",
    "psycopg2-binary>=2.9",
    "duckdb>=1.0",
    "boto3>=1.34",
]

[project.scripts]
yaci-s3 = "yaci_s3.cli:main"

[dependency-groups]
dev = [
    "pytest>=8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/yaci_s3"]
